+++ lib/request_handler/base.rb
@@ -96,11 +96,14 @@ module RequestHandler
 
     def parse_body_params
       defaults = fetch_defaults('body.defaults', {})
-      defaults.merge(BodyParser.new(
-        request:        request,
-        schema:         lookup!('body.schema'),
-        schema_options: execute_options(lookup('body.options'))
-      ).run)
+      parsed_body = BodyParser.new(
+        request:          request,
+        schema:           lookup!('body.schema'),
+        schema_options:   execute_options(lookup('body.options')),
+        included_schemas: lookup('body.included')
+      ).run
+      return defaults.merge(parsed_body) unless parsed_body.is_a? Array
+      parsed_body.unshift(defaults.merge(parsed_body.shift))
     end
 
     def parse_fieldsets_params